feat(SCRUM-91): add get all order for a shop

+++ src/main/java/fr/shoploc/shoplocBackend/productInCart/service/ProductInCartService.java
@@ -1,16 +1,15 @@
 package fr.shoploc.shoplocBackend.productInCart.service;
 
+import fr.shoploc.shoplocBackend.common.methods.Common;
 import fr.shoploc.shoplocBackend.common.models.Product;
 import fr.shoploc.shoplocBackend.common.models.ProductInCart;
 import fr.shoploc.shoplocBackend.common.models.Shop;
-import fr.shoploc.shoplocBackend.config.JwtService;
 import fr.shoploc.shoplocBackend.dto.ProductDTO;
-import fr.shoploc.shoplocBackend.dto.ShopDTO;
+import fr.shoploc.shoplocBackend.dto.CartDTO;
+import fr.shoploc.shoplocBackend.product.repository.ProductRepository;
 import fr.shoploc.shoplocBackend.productInCart.repository.ProductInCartRepository;
 import fr.shoploc.shoplocBackend.product.controller.ProductController;
 import fr.shoploc.shoplocBackend.shop.controller.ShopController;
-import fr.shoploc.shoplocBackend.usermanager.user.User;
-import fr.shoploc.shoplocBackend.usermanager.user.UserService;
 import org.springframework.stereotype.Service;
 
 import java.util.*;
@@ -19,20 +18,20 @@ import java.util.*;
 public class ProductInCartService {
 
     private final ProductInCartRepository productInCartRepository;
+    private final ProductRepository productRepository;
     private final ProductController productController;
-    private final JwtService jwtService;
-    private final UserService userService;
+    private final Common common;
     private final ShopController shopController;
 
-    public ProductInCartService(ProductInCartRepository productInCartRepository, ProductController productController, JwtService jwtService, UserService userService, ShopController shopController){
+    public ProductInCartService(ProductInCartRepository productInCartRepository, ProductRepository productRepository, ProductController productController, Common common, ShopController shopController){
         this.productInCartRepository = productInCartRepository;
+        this.productRepository = productRepository;
         this.productController = productController;
-        this.jwtService = jwtService;
-        this.userService = userService;
+        this.common = common;
         this.shopController = shopController;
     }
     public void addProductToCart(Long idProduct, String token) throws Exception {
-        Long userId = getUserId(token);
+        Long userId = common.getUserId(token);
 
         List<ProductInCart> existingProductInCart = productInCartRepository.findByIdUserAndIdProduct(userId, idProduct);
 
@@ -47,25 +46,48 @@ public class ProductInCartService {
     }
 
     public void removeProductToCart(Long idProduct, String token) throws Exception {
-        Long userId = getUserId(token);
-        productInCartRepository.deleteByIdProductAndIdUser(idProduct, userId);
+        Long userId = common.getUserId(token);
+        List<ProductInCart> existingProductInCart = productInCartRepository.findByIdUserAndIdProduct(userId, idProduct);
+
+        if (!existingProductInCart.isEmpty()) {
+            ProductInCart productInCart = existingProductInCart.get(0);
+            if (productInCart.getQuantity() > 1) {
+                productInCart.setQuantity(productInCart.getQuantity() - 1);
+                productInCartRepository.save(productInCart);
+            } else {
+                productInCartRepository.deleteByIdProductAndIdUser(idProduct, userId);
+            }
+        } else {
+            throw new Exception("Produit non trouvé dans le panier.");
+        }
     }
 
-    public List<ShopDTO> getCarts(String token) throws Exception {
-        Long userId = getUserId(token);
+    public List<CartDTO> getUserCarts(String token) throws Exception {
+        Long userId = common.getUserId(token);
 
         List<ProductInCart> productInCartList = productInCartRepository.findAllByIdUser(userId);
+        return getCarts(productInCartList);
+    }
 
-        Map<Long, ShopDTO> cartsByShop = new HashMap<>();
+    public List<ProductInCart> getShopCarts(String token) throws Exception {
+        Long userId = common.getUserId(token);
+        List<Long> productIds = new ArrayList<>();
+        productRepository.findAllByShopId(userId).forEach(product -> productIds.add(product.getId()));
+
+        return productInCartRepository.findByIdProductIn(productIds);
+    }
+    
+    private List<CartDTO> getCarts(List<ProductInCart> productInCartList) throws Exception {
+        Map<Long, CartDTO> cartsByShop = new HashMap<>();
         for (ProductInCart productInCart : productInCartList) {
             Product product = productController.getProductById(productInCart.getIdProduct());
             Long shopId = product.getShopId();
 
-            if (shopController.getShopById(shopId).isEmpty()) {
+            Optional<Shop> optionalShop = shopController.getShopById(shopId);
+            if (optionalShop.isEmpty()) {
                 throw new Exception("Magasin introuvable.");
             }
-
-            Shop shop = shopController.getShopById(shopId).get();
+            Shop shop = optionalShop.get();
 
             ProductDTO productDTO = new ProductDTO();
             productDTO.setId(product.getId());
@@ -74,27 +96,16 @@ public class ProductInCartService {
             productDTO.setQuantity(productInCart.getQuantity());
             productDTO.setImageUrl(product.getImageUrl());
 
-            ShopDTO shopDTO = cartsByShop.get(shopId);
-            if (shopDTO == null) {
-                shopDTO = new ShopDTO();
-                shopDTO.setId(shop.getId());
-                shopDTO.setShopName(shop.getName());
-                cartsByShop.put(shopId, shopDTO);
+            CartDTO cartDTO = cartsByShop.get(shopId);
+            if (cartDTO == null) {
+                cartDTO = new CartDTO();
+                cartDTO.setId(productInCart.getId());
+                cartDTO.setShopId(shopId);
+                cartDTO.setShopName(shop.getName());
+                cartsByShop.put(shopId, cartDTO);
             }
-            shopDTO.getProducts().add(productDTO);
+            cartDTO.getProducts().add(productDTO);
         }
-
         return new ArrayList<>(cartsByShop.values());
-    }
-
-
-    public Long getUserId(String token) throws Exception {
-        String userEmail = jwtService.extractUserEmail(token);
-        Optional<User> user = userService.findUserByEmail(userEmail);
-        if (user.isPresent()) {
-            return user.get().getId();
-        } else {
-            throw new Exception("Utilisateur introuvable.");
-        }
     }
 }
